// src/api-services/generated/models/paged-options.ts
/* tslint:disable */
/* eslint-disable */
/**
 * merged spec
 * merged spec
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface PagedOptions {
    /**
     * 名称
     */
    'label'?: string | null;
    /**
     * 值
     */
    'value'?: object | null;
    /**
     * 是否禁用
     */
    'disaebled'?: boolean;
    /**
     * 说明
     */
    'desc'?: string | null;
    /**
     * 子元素
     */
    'children'?: Array<PagedOptions> | null;
}